Resolve circular dep in assets/pods
